feat: add coutting with discount

+++ back-end/src/utils/count_goods_prices.py
@@ -4,8 +4,8 @@ from functools import reduce
 def countGoodsPrices(goodsPrices: Union[int, List[int]], discount: Union[int, float] = 0) -> int:
     goodsPricesDiscount = 1 - discount 
     if type(goodsPrices) == int:
-        return goodsPrices * goodsPricesDiscount
+        return int(goodsPrices * goodsPricesDiscount)
     else:
         totalPriceOrigin = reduce(lambda acc, nextV: acc + nextV, goodsPrices, 0)
         totalPriceDiscounted = totalPriceOrigin * goodsPricesDiscount
-        return totalPriceDiscounted
+        return int(totalPriceDiscounted)
